refactor(auth): collapse auth/ decomposition, delegate JWKS, delete WebAuthnSelfAuthVerifier (S3)
Sprint S3 Tasks 1.0+2.0+3.0 — auth/ over-decomposition cleanup:

- Merge claims.py -> jwks.py + middleware.py; utils.py -> scopes.py;
  lifecycle.py -> identity.py (auth/ 16 -> 13 files).
- Trim leaky auth/__init__.py re-exports (Claims/fetch_keys/validate_jwt/
  host_urn_from_thumbprint) — grep-verified zero external root callers.
- OAuth2Middleware now delegates to a shared JWKSValidator (one canonical
  JWKS cache; TTL reconciled to 86400). server.py collapses the two
  OAuth2Middleware instances (S0 #237 deferred landmine) into one shared
  validator across the HTTP stack + WS app.state path (B4/C-2 property).
- Delete WebAuthnSelfAuthVerifier pass-through; WebAuthnVerifierImpl
  satisfies the WebAuthnVerifier protocol directly. Split webauthn.py
  (473 LOC) into webauthn_store.py + webauthn.py (300 LOC).
- agent_jwt.py: drop agent_store_writable mode flag (caller owns
  persistence), extract _validate_standard_claims dedup, drop cast in
  _okp_from_public_jwk. Unify scope constants (capabilities.py imports
  SCOPE_* from scopes; mirrored _SCOPE_* deleted).
- _validate_identity_binding returns a 2-tuple (was an illegible 4-tuple
  with a dead slot + log-only flag).

Behavior-preserving; public surface stable. The scopes<->middleware
cycle is resolved by one documented inline import of parse_scope inside
OAuth2Middleware.validate_bearer_token (no-inline-imports strict-cycle
exception).
