+++ src/asap/auth/__init__.py
@@ -8,7 +8,6 @@ This module provides OAuth2/OIDC authentication for agent-to-agent communication
 Public exports:
     Token: OAuth2 access token model
     OAuth2ClientCredentials: Client for client_credentials grant
-    OAuth2ClientCredentials: Client for client_credentials grant
     OAuth2Middleware: JWT validation middleware (JWKS)
     OAuth2Config: Config for enabling OAuth2 on the server (create_app)
     OAuth2Claims: JWT claims (sub, scope, exp)
@@ -33,28 +32,23 @@ from asap.auth.identity import (
     AgentStore,
     HostIdentity,
     HostStore,
-    host_urn_from_thumbprint,
     jwk_thumbprint_sha256,
 )
 from asap.auth.introspection import TokenInfo, TokenIntrospector
-from asap.auth.jwks import JWKSValidator, Claims, fetch_keys, validate_jwt
-from asap.auth.oidc import OIDCConfig, OIDCDiscovery
+from asap.auth.jwks import JWKSValidator
 from asap.auth.middleware import OAuth2Claims, OAuth2Config, OAuth2Middleware
 from asap.auth.oauth2 import OAuth2ClientCredentials, Token
+from asap.auth.oidc import OIDCConfig, OIDCDiscovery
 from asap.auth.scopes import SCOPE_ADMIN, SCOPE_EXECUTE, SCOPE_READ, require_scope
 
 __all__ = [
     "AgentSession",
     "AgentStore",
-    "Claims",
+    "HostIdentity",
+    "HostStore",
     "JWKSValidator",
     "OIDCConfig",
     "OIDCDiscovery",
-    "fetch_keys",
-    "HostIdentity",
-    "HostStore",
-    "host_urn_from_thumbprint",
-    "jwk_thumbprint_sha256",
     "OAuth2Claims",
     "OAuth2ClientCredentials",
     "OAuth2Config",
@@ -67,8 +61,8 @@ __all__ = [
     "TokenIntrospector",
     "create_agent_jwt",
     "create_host_jwt",
+    "jwk_thumbprint_sha256",
     "require_scope",
-    "validate_jwt",
     "verify_agent_jwt",
     "verify_host_jwt",
 ]
