+++ src/asap/auth/middleware.py
@@ -7,7 +7,6 @@ Supports Custom Claims identity binding (ADR-17).
 
 from __future__ import annotations
 
-import asyncio
 import json
 import os
 import time
@@ -21,9 +20,7 @@ from joserfc import jwk
 from joserfc.errors import JoseError
 from starlette.middleware.base import BaseHTTPMiddleware
 
-from asap.auth.claims import parse_expected_audience_from_env
-from asap.auth.jwks import validate_jwt
-from asap.auth.utils import parse_scope
+from asap.auth.jwks import JWKSValidator
 from asap.observability import get_logger
 
 logger = get_logger(__name__)
@@ -56,6 +53,25 @@ def _parse_subject_map() -> dict[str, str | list[str]]:
     if not isinstance(data, dict):
         return {}
     return {str(k): v for k, v in data.items() if isinstance(v, (str, list))}
+
+
+def parse_expected_audience_from_env(raw: str | None) -> str | list[str] | None:
+    """Parse ``ASAP_AUTH_AUDIENCE`` env value (comma-separated list allowed).
+
+    Example:
+        >>> parse_expected_audience_from_env("urn:asap:agent:identity")
+        'urn:asap:agent:identity'
+        >>> parse_expected_audience_from_env(" a , b , c ")
+        ['a', 'b', 'c']
+    """
+    if not raw or not raw.strip():
+        return None
+    parts = [p.strip() for p in raw.split(",") if p.strip()]
+    if not parts:
+        return None
+    if len(parts) == 1:
+        return parts[0]
+    return parts
 
 
 @dataclass
@@ -114,22 +130,7 @@ def _get_bearer_token(request: Request) -> str | None:
     return auth[7:].strip() or None
 
 
-DEFAULT_HTTP_TIMEOUT = 10.0
-
 AGENT_IDENTITY_ROUTE_PREFIX = "/asap/agent/"
-
-
-async def _fetch_jwks(
-    jwks_uri: str, transport: httpx.AsyncBaseTransport | None = None
-) -> jwk.KeySet:
-    """Fetch JWKS from URI and return a joserfc KeySet."""
-    async with httpx.AsyncClient(
-        transport=transport, timeout=httpx.Timeout(DEFAULT_HTTP_TIMEOUT)
-    ) as client:
-        resp = await client.get(jwks_uri)
-        resp.raise_for_status()
-        data = resp.json()
-    return jwk.KeySet.import_key_set(data)
 
 
 class OAuth2Middleware(BaseHTTPMiddleware):
@@ -155,12 +156,12 @@ class OAuth2Middleware(BaseHTTPMiddleware):
         custom_claim: str | None = None,
         expected_issuer: str | None = None,
         expected_audience: str | list[str] | None = None,
+        validator: JWKSValidator | None = None,
     ) -> None:
         super().__init__(app)
         self._jwks_uri = jwks_uri
         self._required_scope = required_scope
         self._path_prefix = path_prefix
-        self._jwks_fetcher = jwks_fetcher or _fetch_jwks
         self._manifest_id = manifest_id
         self._custom_claim = custom_claim
         self._custom_claim_key = self._custom_claim or os.environ.get(
@@ -171,36 +172,27 @@ class OAuth2Middleware(BaseHTTPMiddleware):
             os.environ.get(ENV_AUDIENCE)
         )
         self._subject_map = _parse_subject_map()
-        self._jwks_cache: jwk.KeySet | None = None
-        self._jwks_cache_time: float = 0.0
-        self._jwks_cache_ttl = 3600.0  # 1 hour
-        self._jwks_lock = asyncio.Lock()
+        # One canonical JWKS cache lives on the shared JWKSValidator so the HTTP
+        # middleware stack instance and the WS ``app.state`` instance do not
+        # diverge (S0 #237). When callers hand us a pre-built validator (server
+        # wiring shares ONE across both transports) we adopt it; otherwise we
+        # build a private one, threading the test ``jwks_fetcher`` seam through.
+        if validator is not None:
+            self._validator = validator
+        else:
+            self._validator = JWKSValidator(
+                jwks_uri,
+                fetcher=jwks_fetcher,
+                expected_issuer=self._expected_issuer,
+                expected_audience=self._expected_audience,
+                require_exp=False,
+            )
 
     def _should_validate(self, path: str) -> bool:
         """Return True if this request path should be validated."""
         if self._path_prefix is not None and not path.startswith(self._path_prefix):
             return False
         return not path.startswith(AGENT_IDENTITY_ROUTE_PREFIX)
-
-    async def _get_key_set(self) -> jwk.KeySet:
-        """Return JWKS, from cache or by fetching."""
-        async with self._jwks_lock:
-            now = time.time()
-            if (
-                self._jwks_cache is not None
-                and (now - self._jwks_cache_time) < self._jwks_cache_ttl
-            ):
-                return self._jwks_cache
-            key_set = await self._jwks_fetcher(self._jwks_uri)
-            self._jwks_cache = key_set
-            self._jwks_cache_time = now
-            return key_set
-
-    async def _invalidate_jwks_cache(self) -> None:
-        """Clear JWKS cache."""
-        async with self._jwks_lock:
-            self._jwks_cache = None
-            self._jwks_cache_time = 0.0
 
     def _validate_scope(self, scope_list: list[str]) -> bool:
         """Check if required_scope is present in scope list."""
@@ -210,10 +202,16 @@ class OAuth2Middleware(BaseHTTPMiddleware):
 
     def _validate_identity_binding(
         self, claims: dict[str, Any], sub: str
-    ) -> tuple[bool, str | None, bool, bool]:
-        """Validate JWT identity binding via custom claim or allowlist (ADR-17)."""
+    ) -> tuple[bool, str | None]:
+        """Validate JWT identity binding via custom claim or allowlist (ADR-17).
+
+        Returns ``(True, None)`` on success and ``(False, detail)`` on mismatch.
+        Allowlist hits are logged here (``identity_via_allowlist``) rather than
+        surfaced as a separate flag to the caller, so the result is a flat
+        ``(ok, error_detail)`` pair.
+        """
         if self._manifest_id is None:
-            return True, None, False, False
+            return True, None
 
         claim_key = self._custom_claim_key
         claim_value = claims.get(claim_key)
@@ -224,39 +222,43 @@ class OAuth2Middleware(BaseHTTPMiddleware):
                 return (
                     False,
                     f"Custom claim {claim_key!r} has {agent_id!r}, expected {self._manifest_id!r}",
-                    False,
-                    False,
                 )
-            return True, None, False, False
+            return True, None
 
         allowed = self._subject_map.get(self._manifest_id)
         if allowed is not None:
             if isinstance(allowed, str) and allowed == sub:
-                return True, None, False, True
+                logger.warning(
+                    "asap.oauth2.identity_via_allowlist",
+                    manifest_id=self._manifest_id,
+                    sub=sub,
+                )
+                return True, None
             if isinstance(allowed, list) and sub in allowed:
-                return True, None, False, True
+                logger.warning(
+                    "asap.oauth2.identity_via_allowlist",
+                    manifest_id=self._manifest_id,
+                    sub=sub,
+                )
+                return True, None
 
-        return False, "Missing required identity claim", True, False
+        return False, "Missing required identity claim"
 
-    async def _validate_token_claims(self, token: str, key_set: jwk.KeySet) -> dict[str, Any]:
+    async def _validate_token_claims(self, token: str) -> dict[str, Any]:
+        """Decode+validate the JWT via the shared validator with key-rotation retry.
+
+        Fetches the JWKS (cached on the validator), decodes the token, and on a
+        ``JoseError`` (e.g. unknown ``kid`` from key rotation) invalidates the
+        cache and retries once. Network errors surface as ``httpx.HTTPError``
+        for the caller to map to a 503 response.
+        """
+        key_set = await self._validator.fetch_keys(self._jwks_uri)
         try:
-            return validate_jwt(
-                token,
-                key_set,
-                expected_issuer=self._expected_issuer,
-                expected_audience=self._expected_audience,
-                require_exp=False,
-            )
+            return self._validator.validate_jwt(token, key_set)
         except JoseError:
-            await self._invalidate_jwks_cache()
-            key_set = await self._get_key_set()
-            return validate_jwt(
-                token,
-                key_set,
-                expected_issuer=self._expected_issuer,
-                expected_audience=self._expected_audience,
-                require_exp=False,
-            )
+            await self._validator.invalidate_cache()
+            key_set = await self._validator.fetch_keys(self._jwks_uri)
+            return self._validator.validate_jwt(token, key_set)
 
     async def validate_bearer_token(
         self, token: str, *, path: str = ""
@@ -269,8 +271,7 @@ class OAuth2Middleware(BaseHTTPMiddleware):
         failure. ``path`` is used only for log context.
         """
         try:
-            key_set = await self._get_key_set()
-            claims = await self._validate_token_claims(token, key_set)
+            claims = await self._validate_token_claims(token)
         except httpx.HTTPError as e:
             logger.error(
                 "asap.oauth2.jwks_fetch_failed",
@@ -326,6 +327,11 @@ class OAuth2Middleware(BaseHTTPMiddleware):
                 headers={"WWW-Authenticate": "Bearer"},
             )
 
+        # Imported here (not at module top) to break the scopes <-> middleware
+        # import cycle: scopes.require_scope references OAuth2Claims at runtime,
+        # so scopes must import middleware, which in turn needs parse_scope.
+        from asap.auth.scopes import parse_scope
+
         scope_list = parse_scope(claims.get("scope"))
         if not self._validate_scope(scope_list):
             logger.warning(
@@ -339,7 +345,7 @@ class OAuth2Middleware(BaseHTTPMiddleware):
                 content={"detail": ERROR_INSUFFICIENT_SCOPE},
             )
 
-        success, err_detail, _, use_allowlist = self._validate_identity_binding(claims, sub)
+        success, err_detail = self._validate_identity_binding(claims, sub)
         if not success:
             logger.warning(
                 "asap.oauth2.identity_mismatch",
@@ -354,13 +360,6 @@ class OAuth2Middleware(BaseHTTPMiddleware):
             return None, JSONResponse(
                 status_code=HTTP_FORBIDDEN,
                 content={"detail": detail_msg},
-            )
-        if use_allowlist:
-            logger.warning(
-                "asap.oauth2.identity_via_allowlist",
-                path=path,
-                manifest_id=self._manifest_id,
-                sub=sub,
             )
 
         return OAuth2Claims(sub=sub, scope=scope_list, exp=exp_ts), None
